requirements.txt
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
passlib==1.7.1
psycopg2==2.8.3
PyJWT==1.4.2
Werkzeug==0.15.6
